DRIVERS-3605: extract otel.py; support OTel via mongodb-runner
- Move OTel constants and helpers into a sibling otel.py module
  (and add sibling modules to the hatch build include list).
- Lift the --mongodb-runner restriction: the runner passes
  procParams.setParameter through; verified live via run-mongodb.sh.
- Point docs at run-mongodb.sh and cover the runner path in test-otel.sh.
