+++ .evergreen/orchestration/pyproject.toml
@@ -27,7 +27,7 @@ dependencies = [
 drivers-orchestration = "drivers_orchestration:main"
 
 [tool.hatch.build]
-include = ["drivers_orchestration.py"]
+include = ["drivers_orchestration.py", "mongodb_runner.py", "otel.py"]
 
 [tool.hatch.metadata]
 allow-direct-references=true
